Add function to create debug temporary file
This commit introduces a function, debug_tmp_file_create_(), in namespace nutsloop of the new file "log.h++". This function is used for creating a unique and secure temporary file if the DEBUG_LOG is enabled. This temporary file will be used for debug logging. To handle any errors during file creation or closing, proper error handling is incorporated. The path of the temporary file is stored safely and atomically for future reference.